backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.3
email-validator>=2.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
